requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
py-slippi==1.6.2
py-ubjson==0.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
termcolor==2.3.0
tzdata==2023.3
